Logic timestamps branch

+++ hds/client/src/main/java/hds/client/domain/IntentionToSellCallable.java
@@ -11,10 +11,10 @@ import java.util.concurrent.Callable;
 
 import static hds.client.helpers.ClientProperties.getMyClientPort;
 import static hds.client.helpers.ClientProperties.getMyPrivateKey;
-import static hds.client.helpers.ConnectionManager.*;
 import static hds.security.ConvertUtils.bytesToBase64String;
 import static hds.security.SecurityManager.newWriteOnGoodsData;
 import static hds.security.SecurityManager.setMessageSignature;
+import static hds.security.helpers.managers.ConnectionManager.*;
 
 /**
  * The type get intention to sell callable performs a PUT request to the end point /stateOfGood of a notary replica
@@ -44,7 +44,7 @@ public class IntentionToSellCallable implements Callable<BasicMessage> {
                                    String requestId,
                                    String replicaId,
                                    String goodId,
-                                   long wts,
+                                   int wts,
                                    Boolean onSale) {
 
         this.replicaId = replicaId;
@@ -81,7 +81,7 @@ public class IntentionToSellCallable implements Callable<BasicMessage> {
      * @throws JSONException
      * @throws SignatureException
      */
-    private byte[] newWriteOnGoodsDataSignature(final String goodId, final Boolean onSale, final long wts)
+    private byte[] newWriteOnGoodsDataSignature(final String goodId, final Boolean onSale, final int wts)
             throws JSONException, SignatureException {
 
         byte[] rawData = newWriteOnGoodsData(goodId, onSale, getMyClientPort(), wts).toString().getBytes();
@@ -105,7 +105,7 @@ public class IntentionToSellCallable implements Callable<BasicMessage> {
                                                  final String requestId,
                                                  final String to,
                                                  final String goodId,
-                                                 final long wts,
+                                                 final int wts,
                                                  final Boolean onSale,
                                                  final String writeOnGoodsSignature) {
