+++ hds/client/src/main/java/hds/client/helpers/ClientSecurityManager.java
@@ -6,11 +6,12 @@ import org.json.JSONException;
 
 import java.security.SignatureException;
 
-import static hds.client.helpers.ClientProperties.*;
 import static hds.client.helpers.ClientProperties.getMyPrivateKey;
+import static hds.client.helpers.ClientProperties.printError;
 import static hds.security.SecurityManager.*;
 
 public class ClientSecurityManager {
+
 
     public static boolean isMessageFreshAndAuthentic(BasicMessage message) {
         // Verify freshness and authenticity using isValidMessage
@@ -29,7 +30,7 @@ public class ClientSecurityManager {
     public static byte[] newWriteOnGoodsDataSignature(final String goodId,
                                                       final Boolean onSale,
                                                       final String writer,
-                                                      final long wts) throws JSONException, SignatureException {
+                                                      final int wts) throws JSONException, SignatureException {
 
         byte[] rawData = newWriteOnGoodsData(goodId, onSale, writer, wts).toString().getBytes();
         return CryptoUtils.signData(getMyPrivateKey(), rawData);
@@ -37,7 +38,7 @@ public class ClientSecurityManager {
 
     public static byte[] newWriteOnOwnershipsDataSignature(final String goodId,
                                                            final String writerID,
-                                                           final long wts) throws JSONException, SignatureException {
+                                                           final int wts) throws JSONException, SignatureException {
 
         byte[] rawData = newWriteOnOwnershipData(goodId, writerID, wts).toString().getBytes();
         return CryptoUtils.signData(getMyPrivateKey(), rawData);
